New upstream release
Deal with some legitimate lintian warnings, and add overrides for the
new things it's barking up a wrong tree about.  That in turns buys us
a new dep on dh > 6.0.7, to silence warnings about problems which didn't
exist before, and still don't exist now.  And so the wheels of progress
grind flat the cobble-stones of history.

And yet another nail in nvi's coffin.  Long live The Vim.