debian: Install config file in /etc/gtags
global includes a configuration file as an example in the
documents. The user is expected to copy this into their $HOME
folder. Let's simplify the process by installing the file as a
system-wide config.

It does not break existing changes for users who have made changes to
their config as the local changes override the system wide defaults.