debian/rules: Adjust configure to enable universal-ctags
Upstream global supports both exuberant and universal ctags but the
debian package only enabled exuberant ctags support. It also assumed
that /usr/local/ctags is pointing to exuberant ctags which is wrong
depending on which version of ctags is installed in the system.

Update the configure rules to explicitly refer to both exuberant and
universal verions of ctags.

[Punit Agrawal]
* Rebase to latest upstream
* Update commit message