debian/control: Add python3-minimal as a build dependency
Recent versions of Global are relying on the python interpreter
available during build to set the appropriate interpreter for pygments
parser. In the absence of any suitable interpreter, the default is to
use "/usr/bin/python" which is not available by default in bookworm.

Add python3-minimal to the build dependencies to correctly discover
the interpreter to use.